feat: fjärrvärmemätning via Landis+Gyr UltraHeat T550
- Vendorar ESPHome-komponenten uh50 (psvanstrom/esphome-uh50reader, MIT)
  som iec-fjarrvarme.yaml använder för att läsa T550-mätaren
- iec62056: castar loggargument explicit så format-specificerarna
  matchar på alla plattformar
- template_sensors: sensor.fjarrvarme_pris med Tekniska verkens
  säsongstariff för 2026 (vinter/vår-höst/sommar) + metadata-attribut
- dashboards/elektricitet: ny vy "Fjärrvärme" med effekt, tillopps- och
  returtemperatur, pris och förbrukningsgraf
- esphome/.gitignore: ignorera __pycache__ från local_components

+++ esphome/local_components/iec62056/iec62056.cpp
@@ -53,7 +53,7 @@ void IEC62056Component::dump_config() {
     // These settings are not used in Mode D
     ESP_LOGCONFIG(TAG, "  Battery meter: %s", YESNO(this->battery_meter_));
     if (this->config_baud_rate_max_bps_ > 0) {
-      ESP_LOGCONFIG(TAG, "  Max baud rate: %u bps", this->config_baud_rate_max_bps_);
+      ESP_LOGCONFIG(TAG, "  Max baud rate: %u bps", (unsigned) this->config_baud_rate_max_bps_);
     } else {
       ESP_LOGCONFIG(TAG, "  Max baud rate: not limited");
     }
@@ -311,7 +311,7 @@ void IEC62056Component::loop() {
           connection_status_(false);
 
           // end of data
-          ESP_LOGD(TAG, "Total connection time: %u ms", millis() - retry_connection_start_timestamp_);
+          ESP_LOGD(TAG, "Total connection time: %u ms", (unsigned) (millis() - retry_connection_start_timestamp_));
 
           verify_all_sensors_got_value_();
           ESP_LOGD(TAG, "Start of sensor update");
@@ -408,7 +408,7 @@ void IEC62056Component::loop() {
         ESP_LOGD(TAG, "Meter reported protocol: %c", (char) mode_);
         if (mode_ != PROTOCOL_MODE_A) {
           ESP_LOGD(TAG, "Meter reported max baud rate: %u bps ('%c')",
-                   identification_to_baud_rate_(baud_rate_identification_), baud_rate_identification_);
+                   (unsigned) identification_to_baud_rate_(baud_rate_identification_), baud_rate_identification_);
         }
         set_next_state_(PREPARE_ACK);
       }
@@ -436,10 +436,10 @@ void IEC62056Component::loop() {
         }
 
         baud_rate_char = baud_rate_to_identification_(negotiated_bps);
-        ESP_LOGD(TAG, "Using negotiated baud rate %d bps.", negotiated_bps);
+        ESP_LOGD(TAG, "Using negotiated baud rate %d bps.", (int) negotiated_bps);
       } else {
         ESP_LOGD(TAG, "Using meter maximum baud rate %d bps ('%c').",
-                 identification_to_baud_rate_(baud_rate_identification_), baud_rate_identification_);
+                 (int) identification_to_baud_rate_(baud_rate_identification_), baud_rate_identification_);
         baud_rate_char = baud_rate_identification_;
       }
 
@@ -450,8 +450,8 @@ void IEC62056Component::loop() {
         } else if (baud_rate_char < PROTO_C_RANGE_BEGIN) {
           baud_rate_char = PROTO_C_RANGE_BEGIN;
         }
-        ESP_LOGD(TAG, "Decreased baud rate for retry %u to: %d bps ('%c').", retry_counter_,
-                 identification_to_baud_rate_(baud_rate_char), baud_rate_char);
+        ESP_LOGD(TAG, "Decreased baud rate for retry %u to: %d bps ('%c').", (unsigned) retry_counter_,
+                 (int) identification_to_baud_rate_(baud_rate_char), baud_rate_char);
       }
 
       data_out_size_ = sizeof(set_baud);
@@ -468,7 +468,7 @@ void IEC62056Component::loop() {
       break;
 
     case SET_BAUD_RATE:
-      ESP_LOGD(TAG, "Switching to new baud rate %u bps ('%c')", new_baudrate, baud_rate_char);
+      ESP_LOGD(TAG, "Switching to new baud rate %u bps ('%c')", (unsigned) new_baudrate, baud_rate_char);
       update_baudrate_(new_baudrate);
       set_next_state_(WAIT_FOR_STX);
       break;
@@ -495,7 +495,7 @@ void IEC62056Component::loop() {
       if ((frame_size = receive_frame_())) {
         // ETX is the first byte (the way receive_frame_() works and \r\n in data)
         if (in_buf_[0] == ETX) {
-          ESP_LOGD(TAG, "Total connection time: %u ms", millis() - retry_connection_start_timestamp_);
+          ESP_LOGD(TAG, "Total connection time: %u ms", (unsigned) (millis() - retry_connection_start_timestamp_));
           // include ETX (but exclude STX)
           lrc_ ^= ETX;  // faster than update_lrc_(in_buf_,1);
           bool bcc_failed = false;
@@ -789,7 +789,8 @@ void IEC62056Component::retry_or_sleep_() {
     wait_next_readout_();
   } else {
     retry_counter_inc_();
-    ESP_LOGD(TAG, "Retry %d of %d. Waiting %u ms before the next try", retry_counter_, max_retries_, retry_delay_);
+    ESP_LOGD(TAG, "Retry %d of %d. Waiting %u ms before the next try", retry_counter_, max_retries_,
+             (unsigned) retry_delay_);
     wait_(retry_delay_, BEGIN);
   }
 }
@@ -829,7 +830,8 @@ void IEC62056Component::wait_next_readout_() {
 
   scheduled_timestamp_set_ = false;
   if (is_periodic_readout_enabled_()) {
-    ESP_LOGD(TAG, "Waiting %u ms for the next scheduled readout (every %u ms).", actual_wait_time, update_interval_ms_);
+    ESP_LOGD(TAG, "Waiting %u ms for the next scheduled readout (every %u ms).", (unsigned) actual_wait_time,
+             (unsigned) update_interval_ms_);
     wait_(actual_wait_time, BEGIN);
   } else {
     // UINT32_MAX means no update, use switch to trigger readout
